Initial screen + Sign up/in (#174)
* Finished Inital Screen UI

The buttons already worked, so I simply had to clean up the UI and make
it so that it is responsive to screen size.

Co-authored-by: Aidan <PossiblyAShrub@users.noreply.github.com>

* Finished UI and Removed Key

I accidentally leaked google-services.json, so I added it into the
.gitignore and also regenerated a key. I am also finished the UI of the
Intial Screen, Sign-In and Sign-Up.

* Done UI and Associated Login Classes

I have finished all the inital screen stuff and only need to do the
test, but right now the tests are being a pain, so I will leave them for
now.

* final fixes

* Fixed Test

Fixed the test so that I could pass it and now I will do a PR. Later on
implemnting the tests for each and every UI element etc.

* Last minute fixes

* Think I Fixed Test

The test passed like once, so let me give it a shot.

* Home navigation extension

* Added s to moodEvents

* @before -> @beforeClass

* Requested Changes

Made the eye in the password toggle in visibility states, changed the
gradle file so that it is back up to date again without the gms solution
attempt.

* Finally Done

I remade tests and it is fianlly done ;w;. No wonder ppl hate tests with
such a passion, the tests actually fail more then my UI itself
does...like how does that make any sense??

* Fixes to Tests

* Added Elavation to Sign Up Button

* Elavation for Sign In

I forgot and I wanted to keep it even for both buttons.

* Update ID's for Images
Signed-off-by: “Taha <“tkamil@ualberta.ca”>

* merge errors

* All the Tests Passed

LETS GOOOOOOOOOOO!

* Fixes

* Test Fixes Again

* Final Pushes Before Merge

I fixed some bugs and tested the features of the intial screen etc.

* More QOL

I fixed some minor bigs with errors and generally made it a better
experince.

* Fixes

Remidied app crashes and fixed it.

---------

Co-authored-by: “Taha <“tkamil@ualberta.ca”>
Co-authored-by: Aidan <PossiblyAShrub@users.noreply.github.com>